recompute: Bail early when conditions are not right
The loop potentially adding follow-on nodes to the recomputeHeap
currently calculates a number of conditions, and then only adds a node
if all of them are satisfied.

Unfortuantely, particularly in a situation where there is a highly
interconnected graph with small computations, this means that a large
majority of the time spent is calculating one of these condetions,
rather than actually doing the core calculations. For example, in a
recent benchmark, the following aggregate times were observed:

recompute:                        23.74s
recompute/(*Node).isStale:        12.83s
recompute/(*node).maybeStabilize:  3.60s
recompute/(*Node).isNecessary:     3.25s

Since each of them is necessary, refactor the loop to bail immediately
when any of the conditions are not met.  Reorder the conditions so
that the most common and/or fastest ones are done first.

After this change, the same benchmark reports the following numbers:

recompute:                        13.16s
recompute/(*Node).isStale:         1.59s
recompute/(*node).maybeStabilize:  4.64s
recompute/(*Node).isNecessary:     0.29s

NB that the loops are actually identical; I thought about refactoring
it to make the whole thing more DRY, but figured that it was intended
in the future to look for opportunities to increase the parallelism;
so I've left it the way it is for now.

Signed-off-by: George Dunlap <dunlapg@umich.edu>