recompute_heap_list: Get rid of the map
The recomputeHeapList structure involves both a classic "linked list",
and also a map.  The map would seem to involve a significant speed-up
to two operations: .len(), and .has().  The former is primarily used
to determine whether the entire heap is empty or not (i.e., comparing
.len() to zero); the latter is exclusively used during testing.

Unfortunately, on a graph with a large number of nodes, a
non-negligible amount of time is spent doing operations on this map.
A recent benchmark profile run had the following sample breakdown,
with nearly a quarter of the time spent on adding and removing things
from the heaplist:

.Stabilize:                          15.3s
.Stabilize/.recompute/.addNodeUnsafe: 1.71s
.Stabilize/removeMinUnsafe:           2.14s

Get rid of the map, and use linked list operations instead.

For simplicity and clarity, implement an isEmpty() method, which will
check that the map is non-nil, and that the head is non-nil.  Use this
in all the places where it's required.

In removeMinHeightIter(), we need to move the length check before the
clearing of the heap contents.

After this change, time maintaining the recomputeHeap is significantly
reduced.

This makes the .len() and .has() methods take O(N) to complete.  The
tests are the only thing which use the has() method, and they still
run quite quickly.  The .len() method hasn't yet shown up in the
benchmark results, but it would be easy to optimize by adding a .count
field to the recoomputeHeapList and keeping track of that.

Signed-off-by: George Dunlap <dunlapg@umich.edu>