recompute_heap_list: Don't returning id from pop
The .pop() method seems to take a larger portion of the time than it
should, although it's mainly doing pointer manipulation.  One theory
was that copying Identifier was contributing some of the unexpected
overhead.  The only place pop() is called outside of tests, the
identifier is ignored.

Modify pop() to return only the node and presence, remvoing all
references to it in the test code.

Signed-off-by: George Dunlap <dunlapg@umich.edu>